requirements.txt
fastapi
uvicorn[standard]
python-dotenv
openai
jinja2
chromadb
streamlit
requests
langchain
langgraph
pydantic